Split-step generator improved
Now input signal symmetric with respect to 0. The length of the sequence must be set odd, I set it to 2049, I suggest not changing this value for 2d data. The appropriate parameters for training on 2D data available in the test_spst_gen.py file, we need to transfer them to the config. Number of blocks changed to 32.